Redid order of process
Before, bootstrapping was done on a sample, then the cuts gradually got
smaller. Now, the cuts are made, and bootstrap sampling is done to get
the redshift for that cut. Then the cuts shrink, and bootstrapping is
done again, etc. This way is more reliable, I think. The specific cuts
still need to be tailored, though.