Started from images, rearranges
Main thing is that I rearranged all the files. I am starting from
scratch, since I need to change a lot of things for this iteration of
my code.

I also started on the process of putting images through SExtractor. It
is not functional at all now, I just wanted to commit the movement of
files, which I didn't do until I had already started on this part.