working r-z
I added the old machinery for the actual process of redshift finding. I
changed some things to make it work with some new aspects of the code,
but algorithmically it is the same. I still do not trust my SDSS
calibrations, since some of my clusters are way off from others in the
CMD. The code from catalogs onward seems solid, though.