started resources file, cleaned up globals
I started the creation of a resources pickle file, which will hold data
that the code creates, but doesn’t want to throw away. It will
consolidate a lot of the other files I have floating around.
The globals file was organized.
The ability to search multiple directories for images and catalogs was
added.
The initial redshift plot was added to the rest of the plots, rather
than being saved as a separate file.