Finished process of calculating slopes
Finished the calculation of the red sequence slope in various bands.
Also corrected the file imports to match new file structure.

I’m committing in preparation of a major rehaul that will change the
way a lot of things are done in terms of flexibility for multiple color
combinations. A lot of color things are hardcoded now, and I’m going to
get away from that. I’ll make a new branch for that.